+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## Unreleased
+- Redesigned Skill Page
+
 ## 0.20.0 - Sidebar image and Mobile app link
 - Limit user image size to 50px in the sidebar
 - Added link to download mobile app